requirements.txt

fastapi==0.75.2
Flask==2.1.1

pytesseract==0.3.9

tesseract==0.1.3

waitress==2.1.1
mysql-connector-python==8.0.29

